pfz_backend/requirements.txt
# requirements.txt
# Add your dependencies here
fastapi
uvicorn
pandas
python-multipart
Pillow
# tensorflow  # Disabled for demo - uncomment for production ML model
# numpy       # Disabled for demo - uncomment for production ML model
# scikit-learn
